requirements.txt
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==11.2.1
pypugjs==5.11.0
requests==2.32.3
six==1.16.0
urllib3==2.4.0
uWSGI==2.0.29
Werkzeug==3.0.3
